refactor: rename style_regression to PascalCase StyleRegression
Match the multifactor directory naming convention where
compiled translation units use PascalCase (MultiFactorBase,
NormalizeBase, ScoreRecord) rather than snake_case.
