+++ hikyuu_cpp/hikyuu/trade_sys/multifactor/StyleRegression.h
